Sort channel members in UID order
Previously, the ordering of members in a channel was based on the
reverse ordering in which they joined (i.e. most recent first). While
this worked just fine, the lack of a consistent ordering means that
things like NAMES and WHO output are arbitrary and (more importantly)
that it was impossible to implement asynchronous METADATA listings.

Instead, make the channel member list a radixtree sorted by UID. This
gives consistent iteration ordering while keeping a stable identifier
(so the tree doesn't need to be rebalanced/adjusted on nick changes).
Because radixtrees were previously always exposed in /STATS B and there
is now one radixtree per channel, also add a capability to make an
"anonymous" radixtree that is not exposed to /STATS B and use it for
these new trees.

As part of this, a handful of cases that duplicated the code to clear a
user's accept list were replaced by the helper function that did the
same.

+++ librb/include/rb_radixtree.h
@@ -150,7 +150,7 @@ const char *rb_radixtree_elem_get_key(rb_radixtree_leaf *elem);
 void rb_radixtree_elem_set_data(rb_radixtree_leaf *elem, void *data);
 void *rb_radixtree_elem_get_data(rb_radixtree_leaf *elem);
 
-unsigned int rb_radixtree_size(rb_radixtree *dict);
+size_t rb_radixtree_size(rb_radixtree *dict);
 void rb_radixtree_stats(rb_radixtree *dict, void (*cb)(const char *line, void *privdata), void *privdata);
 void rb_radixtree_stats_walk(void (*cb)(const char *line, void *privdata), void *privdata);
